pyproject.toml
[tool.poetry]
name = "reddit-img-parser"
version = "0.7.0"
description = ""
authors = ["skip"]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = "^2.28.2"
progress = "^1.6"
fake-useragent = "^1.1.1"
flake8 = "^6.0.0"
redgifs = { git = "https://github.com/scrazzz/redgifs.git" }
bs4 = "^0.0.1"
argparse = "^1.4.0"
praw = "^7.6.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
reddit-img-parser = "reddit_img_parser.scripts.reddit_img_parser:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
